Scopes in Bean Declaration
Scopes define how bean will be created. For singleton, it will create single instance and will cache the instance and will use same instance. We can verify that if we put break point in postConstruct method and then debug the application. We will see that projectRepository and projectRepository2 refers to same instance.