Spring Auto configuration -
Spring Boot relies on '@Conditional' annotation and a number of variations to drive auto-configuration.
@ConditionalOnClass
@ConditionalOnMissingClass
@ConditionalOnBean,
@ConditionalOnMissingBean

For example, @ConditionalOnMissingBean will disable a bean and our bean will take priority if defined and found in the class path. We can set 'org.springframework.boot.autoconfigure.logging' to DEBUG mode to get report that contains information on positive matches, negative matches, exclusions and unconditional classes for autoconfiguration.