Request Parameters in URLs -
Spring supports for request parameters using '@RequestParam' annotation. They can forexample, help filter output based on the value of a parameter. This annotation supports 'required' and 'defaultValue' attribute. The required attribute is true by default. If we don't provide a value for the required parameter the application will retrieve an error response.