Spring AOP -
Main components of AOP include:
1. Aspect: actual cross-cutting logic that we want to add
2. Join Point: the point during execution of a program where we can hook in our logic to run
3. Point Cut: a way to match one or more join points and determine where the logic is going to actually run.
4. Advice: this is infrastructure around a join point that runs the Aspect at the Join point.

AOP support is part of basic web and data starters. Here, we have implemented loggin. We want to log the request whenever ProjectServiceImpl.findById() method is executed. To register the class as an aspect, we need to annotate the class with '@Aspect' and make it Spring Component. The join point is the 'findById' method in 'ProjectServiceImpl' class. Various types of advice include before, after and around. Here we used before using '@Before' annotation. The parameter joinPoint is optional, depends on if we need any information from JoinPoint. A pointcut is a Predicate that is written as an expression to match the JoinPoint with the advice. This is what defines the logic linkging with our findById method. Now, when we access http://localhost:8080/projects/1 we'll see logs from Aspect.